react-djang-back/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
da-vinci==0.3.0
Deprecated==1.2.13
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-swagger==2.2.0
django-thumbnails==0.6.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
Pygments==2.10.0
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
